CI Improvements: - Install node environment (node 8.5.0) - Run typscripting linting - Run UI unit tests - Fix issue with UI unit tests - Upload dsub local provider test artifcats

+++ servers/dsub/jobs/test/test_jobs_controller_local.py
@@ -35,7 +35,8 @@ class TestJobsControllerLocal(BaseTestCases.JobsControllerTestCase):
         os.mkdir(self.log_path)
 
     def tearDown(self):
-        shutil.rmtree(self.testing_root)
+        if os.environ.get('CIRCLECI') != 'true':
+            shutil.rmtree(self.testing_root)
         tempfile.tempdir = None
 
     def create_app(self):
